requirements.txt
Django>=4.2.0
django-allauth>=0.57.0
Pillow>=10.0.0
python-decouple>=3.8
